+++ collectors/DNSTrails.py
@@ -10,7 +10,7 @@ def init(domain):
 
 	parser = RawConfigParser()
 	parser.read("config.ini")
-	DNSTrails_API_KEY = parser.get("DNSTrails", "DNSTrails_API_KEY")
+	DNSTrails_API_KEY = parser.get("DNSTrails", "DNSTRAILS_API_KEY")
 
 	if DNSTrails_API_KEY == "":
 		print("  \__", colored("No DNSTrails API key configured", "red"))