+++ CHANGELOG.md
@@ -8,11 +8,14 @@
 ### --- Bug Fixes
 1. Fixed an issue on the wildcard identification logic
 2. Fixed a logic bug on the CertSpotter collector
+3. Replaced Findsubdomains collector with Spyse API
 
 ### --- Misc
 1. Updated README.md
 2. Updated requirements.txt
 3. Updated .gitignore
+4. Added new takeover signatures
+5. Added various records to the words.txt list used for permutations
 
 # Version 3.1.0